+++ frontend-ticketx88/src/pages/Register/Register.jsx
@@ -1,5 +1,5 @@
 import React from "react"
-import SignIn from "./style.js"
+import SignIn from "./signin.js"
 
 const Register = () => {
     return (
