--- frontend-ticketx88/src/pages/Register/style.css
@@ -1,13 +1,0 @@
-.sign-up {
-    background-color: #fff;
-    border: 1px solid #e6ebf5;
-    border-radius: 8px;
-    box-sizing: border-box;
-    height: 390px;
-    width: 360px;
-    padding: 16px;
-    text-size-adjust: 100%;
-    position: relative;
-    margin: 0 auto;
-    box-shadow: 0 1px 3px rgba(0, 0, 0, 0.5);
-}
