+++ frontend-ticketx88/src/styles/Register.css
@@ -1,0 +1,54 @@
+.sign-up {
+    background-color: #fff;
+    border: 1px solid #e6ebf5;
+    border-radius: 8px;
+    box-sizing: border-box;
+    height: 390px;
+    width: 360px;
+    padding: 16px;
+    text-size-adjust: 100%;
+    position: relative;
+    margin: 0 auto;
+    box-shadow: 0 1px 3px rgba(0, 0, 0, 0.5);
+}
+
+.Continue-button {
+    border-radius: 20px;
+    height: 40px;
+    font-weight: bold;
+    font-size: 16px;
+    line-height: 24px;
+    min-width: 88px;
+    padding: 0px 24px;
+    width: 328px;
+    margin-top: 144px;
+    background: rgb(45, 194, 117);
+    color: rgb(255, 255, 255);
+}
+
+.Continue-button:disabled {
+    background: rgb(221, 221, 227);
+    color: rgb(128, 128, 137);
+    cursor: not-allowed;
+}
+
+.checkbox {
+    text-size-adjust: 100%;
+
+}
+
+.checkbox::after {
+
+    background-color:#2dc275;
+border:1px solid #2dc275;
+border-radius:4px;
+bottom:-19px;
+box-sizing:border-box;
+height:16px;
+left:0;
+position:absolute;
+right:310px;
+text-size-adjust:100%;
+top:3px;
+width:16px
+}
